Remove non-applicable test cases
