+++ agglayer/types_test.go
@@ -42,11 +42,11 @@ func TestBridgeExitHash(t *testing.T) {
 		bridge.Hash().String(), "metadata is a hashed and it's empty,use it")
 }
 
-func TestMGenericPPError(t *testing.T) {
+func TestGenericError_Error(t *testing.T) {
 	t.Parallel()
 
-	err := GenericPPError{"test", "value"}
-	require.Equal(t, "Generic error: test: value", err.Error())
+	err := GenericError{"test", "value"}
+	require.Equal(t, "[Agglayer Error] test: value", err.Error())
 }
 
 func TestCertificateHeaderID(t *testing.T) {
@@ -437,15 +437,36 @@ func TestGlobalIndex_UnmarshalFromMap(t *testing.T) {
 
 func TestUnmarshalCertificateHeaderUnknownError(t *testing.T) {
 	t.Parallel()
+	rawCertificateHeader := `{
+		"network_id": 14,
+		"height": 0,
+		"epoch_number": null,
+		"certificate_index": null,
+		"certificate_id": "0x3af88c9ca106822bd141fdc680dcb888f4e9d4997fad1645ba3d5d747059eb32",
+		"new_local_exit_root": "0x625e889ced3c31277c6653229096374d396a2fd3564a8894aaad2ff935d2fc8c",
+		"metadata": "0x0000000000000000000000000000000000000000000000000000000000002f3d",
+		"status": {
+			"InError": {
+				"error": {
+					"ProofVerificationFailed": {
+						"Plonk": "the verifying key does not match the inner plonk bn254 proof's committed verifying key"
+					}
+				}
+			}
+		}
+	}`
 
-	str := "{\"network_id\":14,\"height\":0,\"epoch_number\":null,\"certificate_index\":null,\"certificate_id\":\"0x3af88c9ca106822bd141fdc680dcb888f4e9d4997fad1645ba3d5d747059eb32\",\"new_local_exit_root\":\"0x625e889ced3c31277c6653229096374d396a2fd3564a8894aaad2ff935d2fc8c\",\"metadata\":\"0x0000000000000000000000000000000000000000000000000000000000002f3d\",\"status\":{\"InError\":{\"error\":{\"ProofVerificationFailed\":{\"Plonk\":\"the verifying key does not match the inner plonk bn254 proof's committed verifying key\"}}}}}"
-	data := []byte(str)
 	var result *CertificateHeader
-	err := json.Unmarshal(data, &result)
+	err := json.Unmarshal([]byte(rawCertificateHeader), &result)
 	require.NoError(t, err)
 	require.NotNil(t, result)
-	ppError := result.Error
-	require.Equal(t, `Generic error: ProofVerificationFailed: {"Plonk":"the verifying key does not match the inner plonk bn254 proof's committed verifying key"}`, ppError)
+
+	expectedErr := &GenericError{
+		Key:   "ProofVerificationFailed",
+		Value: "{\"Plonk\":\"the verifying key does not match the inner plonk bn254 proof's committed verifying key\"}",
+	}
+
+	require.Equal(t, expectedErr, result.Error)
 }
 
 func generateTestProof(t *testing.T) types.Proof {
